Add step activation/deactivation feature
It's now possible to directly activate/deactivate a given step (for most sequencers) with a simple click on the corresponding encoder button. Setting the length of a given sequence is now set with a double-click. The index visualisation has been updated to show both the sequence length and the activated and deactivated steps.